Extend kernel-initramfs installed modules
When building a kernel having an initramfs through genkernel, a default list of kernel modules
is installed from a system-installed file (cannot recall the path). Genkernel itself uses
AMODULES_group and MODULES_group to modify kernel modules installed into the initramfs (from genkernel.conf?).

The problem is that it is unknown how to extend this list from Catalyst.
But the option that helped me here is to install all modules. What is "all"
modules, exactly? It is all modules specified from the related kernel Config
file - specified in the Catalyst spec file (ie sysresccd-krnl-stage2-rescue64.spec)

all of them that are specified by the associated kernel config file.